resources/app/requirements.txt
uuid
datetime
openpyxl
comtypes
eel
email
python-pptx
python-dotenv
pytrovich
pymorphy2